fix of usb copy
